Remove unneeded css and enhance UX
Removed the -webkit line from the input element.

Styled the button so that the user now knows that
they are hovering over the button.

Styled the button to be able to look like it is
being pushed down when clicked on.

Input styled so that the user knows that it is
hovering over the correct input.